% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cmip5Var2Stn.R
\name{Cmip5Var2Stn}
\alias{Cmip5Var2Stn}
\title{Cmip5Var2Stn}
\usage{
Cmip5Var2Stn(stnid, varnms, rcpnm, wdir, mdlnm)
}
\arguments{
\item{stnid}{Weather station ID}

\item{varnms}{variable names}

\item{rcpnm}{RCP senario name}

\item{wdir}{Working directory}

\item{mdlnm}{GCM model name}
}
\value{
data  station-based data which contains "year", "month", "yearmon", "date", "prcp", "tmax", "tmin", "wspd", "rhum", "rsds" column names
}
\description{
Read varialbe output files produced by Quantile mapping procedure and combine into station files
}
