% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoQmap.R
\name{DoQmap}
\alias{DoQmap}
\title{DoQmap}
\usage{
DoQmap(qmf, rcpdata, varnm, monid, sdate, edate)
}
\arguments{
\item{qmf}{Quantile mapping information created from historical period using GetQmapFit function}

\item{rcpdata}{GCM modeled data before bias-correction}

\item{varnm}{Variable name among "prcp", "tmax", "tmin", "wspd", "rhum", "rsds"}

\item{monid}{Month in integer}

\item{sdate}{Starting date for applyig quantile mapping}

\item{edate}{Ending date for applyig quantile mapping}
}
\value{
rcpadj Bias-corrected output data using quantile mapping
}
\description{
Apply quantile mapping information obtained from historical period to the future period
}
