% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractPointValue.R
\name{ExtractPointValue}
\alias{ExtractPointValue}
\title{ExtractPointValue}
\usage{
ExtractPointValue(xin, yin, ncfile, wdir)
}
\arguments{
\item{xin}{Longitude in decimal degree}

\item{yin}{Latitue in decimal degree}

\item{ncfile}{File name of NetCDF file}

\item{wdir}{Directory path which contatins NetCDF files}
}
\value{
var Original values extracted from NetCDF files
}
\description{
Read NetCDF file and extract time-series of grid values based on latitude and longitude information
}
