% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetQmapFit.R
\name{GetQmapFit}
\alias{GetQmapFit}
\title{GetQmapFit}
\usage{
GetQmapFit(mrgdata, varnm, monid)
}
\arguments{
\item{mrgdata}{Merged data which contains both observed and scenario data for the common period time}

\item{varnm}{Variable name among "prcp", "tmax", "tmin", "wspd", "rhum", "rsds"}

\item{monid}{Month in ingeger}
}
\value{
qmf Fitting information by quantile mapping
}
\description{
Get Quantile Mapping Fitting information when CommonPeriod option is True
}
