% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetQmapFit2.R
\name{GetQmapFit2}
\alias{GetQmapFit2}
\title{GetQmapFit2}
\usage{
GetQmapFit2(obsdata, rcpdata, varnm, monid)
}
\arguments{
\item{obsdata}{Observed data}

\item{rcpdata}{Scenario data}

\item{varnm}{name among "prcp", "tmax", "tmin", "wspd", "rhum", "rsds"}

\item{monid}{Month in integer}
}
\value{
qmf Fitting information by quantile mapping
}
\description{
Get Quantile Mapping Fitting information when CommonPeriod option is Fale
}
