% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadCmip5DataFromAdss.R
\name{LoadCmip5DataFromAdss}
\alias{LoadCmip5DataFromAdss}
\title{Download clipped national level CMIP5 data}
\usage{
LoadCmip5DataFromAdss(dbdir, NtlCode)
}
\arguments{
\item{dbdir}{directory where downloaded data located}

\item{NtlCode}{2 digit country(national) code}
}
\description{
Download clipped national level CMIP5 data from ADSS`s ftp server. Apec Climate Center Data Service System
}
\examples{
\dontrun{
rSQMSampleProject() 
SetWorkingEnvironment(envfile = "rSQM.yaml")
LoadCmip5DataFromAdss(dbdir = EnvList$dbdir, NtlCode = EnvList$NtlCode)
## You can find the national code at www.apcc21.org 
}
}
