% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumberOfDays.R
\name{NumberOfDays}
\alias{NumberOfDays}
\title{NumberOfDays}
\usage{
NumberOfDays(date)
}
\arguments{
\item{date}{Input date}
}
\value{
nday Number of days within the given month
}
\description{
Decide number of days within the given month (date)
}
