% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObsDataSummary.R
\name{ObsDataSummary}
\alias{ObsDataSummary}
\title{ObsDataSummary}
\usage{
ObsDataSummary(obsdir, stnfile, VarNames, syear_obs, eyear_obs)
}
\arguments{
\item{obsdir}{directory path for station information file}

\item{stnfile}{file name for station information}

\item{VarNames}{string vector contains variables names}

\item{syear_obs}{start year of observation data}

\item{eyear_obs}{end year of observation data}
}
\description{
ObsDataSummary
}
