% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadNetCDF4.R
\name{ReadNetCDF4}
\alias{ReadNetCDF4}
\title{ReadNetCDF4}
\usage{
ReadNetCDF4(wdir, ncfile)
}
\arguments{
\item{wdir}{Work directory path}

\item{ncfile}{NetCDF file name}
}
\value{
outList Output list which includes longitude, latitude, variable, and NC file information
}
\description{
Read NetCDF file and extract necessary information
}
