% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadObsData.R
\name{ReadObsData}
\alias{ReadObsData}
\title{ReadObsData}
\usage{
ReadObsData(stnid, wdir)
}
\arguments{
\item{stnid}{station identification code}

\item{wdir}{Working directory}
}
\value{
data daily time series data after adding "yearmon" and "date" columns and delete "day" column. The column order of output file is ""year", "month", "yearmon", "date", "prcp", "tmax", "tmin", "wspd", "rhum", "rsds"
}
\description{
Read observed data and add column of yearmon, date columns. The column order of input file should be "year", "month", "day", "Precipitation(mm)", "Max. Temperature (C)", "Min. Temperature (C)", "Wind speed(m/s)", "Relative Humidity (fr)",  "Solar Radiation (MJ/m2)".
}
