% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSQMSampleProject.R
\name{rSQMSampleProject}
\alias{rSQMSampleProject}
\title{function to create simple environment for package tutorial}
\usage{
rSQMSampleProject(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. If false, you can skip the verbose description for the structure}
}
\description{
Creates project directories in user's current working directory and load sample project data into proper path for user to perform simple tutorial.
}
\examples{
library(rSQM)
rSQMSampleProject()
## Check your working directory, you can see newly created directories and sample data being loaded.
}
