% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetWorkingEnvironment.R
\name{SetWorkingEnvironment}
\alias{SetWorkingEnvironment}
\title{Setting project environment with external yaml file}
\usage{
SetWorkingEnvironment(envfile, override = list())
}
\arguments{
\item{envfile}{yaml file including environmental infomation}

\item{override}{override}
}
\value{
a list object containing all environmental arguments
}
\description{
User needs quite many arguments throughtout the process. For accomodation, one can write down yaml file with environmental arguments and by using it, everythig is done.
}
\examples{

rSQMSampleProject() 
SetWorkingEnvironment(envfile = "rSQM.yaml")

\dontrun{
EnvList <- SetWorkingEnvironment(envfile = "UserCustomized.yaml")
## And this EnvList object has all the information for the all the next steps
}

}
