% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GhcnDailyUpdate.R
\name{GhcnDailyUpdate}
\alias{GhcnDailyUpdate}
\title{GhcnDailyUpdate}
\usage{
GhcnDailyUpdate(NtlCode, ghcndir, syear_obs, eyear_obs)
}
\arguments{
\item{NtlCode}{2 digit country(national) code, EnvList$NtlCode}

\item{ghcndir}{directory where downloaded data located, EnvList$ghcndir}

\item{syear_obs}{start year of observation, EnvList$syear_obs}

\item{eyear_obs}{end year of observation, EnvList$eyear_obs}
}
\description{
Update Observation Data from GHCN(Global Historical Climatology Network)
}
