% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadCmip5DataFromAdss.R
\name{LoadCmip5DataFromAdss}
\alias{LoadCmip5DataFromAdss}
\title{LoadCmip5DataFromAdss}
\usage{
LoadCmip5DataFromAdss(dbdir, NtlCode)
}
\arguments{
\item{dbdir}{directory where downloaded data located}

\item{NtlCode}{2 digit country(national) code}
}
\description{
Download clipped national level CMIP5 data from APCC's ADSS
Apec Climate Center Data Service System
}
