% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSQMSampleProject.R
\name{rSQMSampleProject}
\alias{rSQMSampleProject}
\title{rSQMSampleProject}
\usage{
rSQMSampleProject()
}
\description{
Build environment settings, create necessary directories and load sample data.
}
