% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetWorkingEnvironment.R
\name{SetWorkingEnvironment}
\alias{SetWorkingEnvironment}
\title{SetWorkingEnvironment}
\usage{
SetWorkingEnvironment(envfile, override = list())
}
\arguments{
\item{envfile}{yaml file including environmental infomation}

\item{override}{override}
}
\description{
Update
}
